+++ Clicked/Config/Addon.lua
@@ -83,13 +83,11 @@ function AddonOptions:CreateOptionsTable()
 				set = function(_, val)
 					Addon.db.profile.options.onKeyDown = val
 
-					Addon:UpdateMacroFrameHandlerPressType()
-
 					for _, frame in Clicked:IterateClickCastFrames() do
-						Clicked:RegisterFrameClicks(frame, true)
+						Clicked:RegisterFrameClicks(frame)
 					end
 
-					Clicked:RegisterFrameClicks(_G[Addon.MACRO_FRAME_HANDLER_NAME], false)
+					Clicked:RegisterFrameClicks(_G[Addon.MACRO_FRAME_HANDLER_NAME])
 					Clicked:ProcessActiveBindings()
 
 					Addon:ShowInformationPopup(Addon.L["If you are using custom unit frames you may have to adjust a setting within the unit frame configuration panel to enable support for this, and potentially even a UI reload."])