+++ CHANGELOG.md
@@ -8,6 +8,12 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 * The MINOR component is used whenever a version has backwards-compatible profile changes. This also indicates that the user can not switch back to a previous MINOR version without using a backup.
 * The PATCH component is used for versions that do not contain profile format changes. Users can freely switch between PATCH versions without risk of data loss.
 
+## [Unreleased]
+
+### Fixed
+
+* Fix more issues with tooltips [#293]
+
 ## [1.17.4] - 2026-02-09
 
 ### Added
@@ -1852,6 +1858,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 [0.4.0]: https://github.com/Snakybo/Clicked/releases/tag/0.4.0
 [0.3.0]: https://github.com/Snakybo/Clicked/releases/tag/0.3.0
 
+[#293]: https://github.com/Snakybo/Clicked/issues/293
 [#291]: https://github.com/Snakybo/Clicked/issues/291
 [#288]: https://github.com/Snakybo/Clicked/issues/288
 [#287]: https://github.com/Snakybo/Clicked/issues/287