+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 * The MINOR component is used whenever a version has backwards-compatible profile changes. This also indicates that the user can not switch back to a previous MINOR version without using a backup.
 * The PATCH component is used for versions that do not contain profile format changes. Users can freely switch between PATCH versions without risk of data loss.
 
-## [Unreleased]
+## [1.17.5] - 2026-02-26
 
 ### Added
 
@@ -1722,7 +1722,8 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 
 * Initial public release
 
-[Unreleased]: https://github.com/Snakybo/Clicked/compare/1.17.4...master
+[Unreleased]: https://github.com/Snakybo/Clicked/compare/1.17.5...master
+[1.17.5]: https://github.com/Snakybo/Clicked/releases/tag/1.17.5
 [1.17.4]: https://github.com/Snakybo/Clicked/releases/tag/1.17.4
 [1.17.3]: https://github.com/Snakybo/Clicked/releases/tag/1.17.3
 [1.17.2]: https://github.com/Snakybo/Clicked/releases/tag/1.17.2